refactor collect_gallery_files, check all file exts
