+++ sphinx_gallery/tests/test_gen_gallery.py
@@ -12,17 +12,20 @@ import pytest
 from sphinx.config import is_serializable
 from sphinx.errors import ConfigError, ExtensionError, SphinxWarning
 from sphinx_gallery.gen_gallery import (
-    check_duplicate_filenames,
-    check_spaces_in_filenames,
-    collect_gallery_files,
     write_computation_times,
     _fill_gallery_conf_defaults,
     write_api_entry_usage,
     fill_gallery_conf_defaults,
+    # check_duplicate_filenames,
+    # check_spaces_in_filenames,
 )
 from sphinx_gallery.interactive_example import create_jupyterlite_contents
-from sphinx_gallery.utils import _escape_ansi
-
+from sphinx_gallery.utils import (
+    _collect_gallery_files,
+    _escape_ansi,
+    check_duplicate_filenames,
+    check_spaces_in_filenames,
+)
 
 MINIMAL_HEADER = """
 '''
@@ -236,7 +239,7 @@ def test_spaces_in_files_warn(sphinx_app_wrapper):
 
     files = ["./a/file1.py", "./a/file2.py", "./a/file 3.py"]
     msg = (
-        "Example file name(s) with space(s) found. Having space(s) in "
+        "Example file name(s) with spaces found. Having spaces in "
         "file names will break some links. "
         "List of files: {}"
     )
@@ -333,8 +336,8 @@ def test_collect_gallery_files(tmpdir, gallery_conf):
         "examples/folder1/file1.py",
         "examples/folder1/file2.py",
         "examples/folder2/file1.py",
-        "tutorials/folder1/subfolder/file1.py",
-        "tutorials/folder2/subfolder/subsubfolder/file1.py",
+        "tutorials/folder1/file1.py",
+        "tutorials/folder2/file1.py",
     ]
 
     abs_paths = [tmpdir.join(rp) for rp in rel_filepaths]
@@ -343,7 +346,9 @@ def test_collect_gallery_files(tmpdir, gallery_conf):
 
     examples_path = tmpdir.join("examples")
     dirs = [examples_path.strpath]
-    collected_files = set(collect_gallery_files(dirs, gallery_conf))
+    collected_files = set(
+        _collect_gallery_files(dirs, gallery_conf, check_filenames=True)
+    )
     expected_files = {
         ap.strpath for ap in abs_paths if re.search(r"examples.*\.py$", ap.strpath)
     }
@@ -352,7 +357,9 @@ def test_collect_gallery_files(tmpdir, gallery_conf):
 
     tutorials_path = tmpdir.join("tutorials")
     dirs = [examples_path.strpath, tutorials_path.strpath]
-    collected_files = set(collect_gallery_files(dirs, gallery_conf))
+    collected_files = set(
+        _collect_gallery_files(dirs, gallery_conf, check_filenames=True)
+    )
     expected_files = {
         ap.strpath for ap in abs_paths if re.search(r".*\.py$", ap.strpath)
     }
@@ -376,7 +383,7 @@ def test_collect_gallery_files_ignore_pattern(tmpdir, gallery_conf):
     gallery_conf["ignore_pattern"] = r"one"
     examples_path = tmpdir.join("examples")
     dirs = [examples_path.strpath]
-    collected_files = set(collect_gallery_files(dirs, gallery_conf))
+    collected_files = set(_collect_gallery_files(dirs, gallery_conf, check_filenames=True))
     expected_files = {
         ap.strpath
         for ap in abs_paths
